requirements_lock.txt
alabaster==0.7.12
astroid==2.9.3
attrs==21.4.0
Babel==2.9.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-default-group==1.2.2
contexttimer==0.3.3
coverage==6.3.1
deprecation==2.1.0
docutils==0.17.1
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
graphviz==0.19.1
idna==3.3
imagesize==1.3.0
immutablecollections==0.12.0
importlib-metadata==4.2.0
incremental==21.3.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.12.0
mypy==0.931
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.6
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
py-cpuinfo==8.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.12.2
pyparsing==3.0.7
pytest==7.0.0
pytest-benchmark==3.4.1
pytest-cov==3.0.0
pytest-pythonpath==0.7.3
pytz==2021.3
PyYAML==6.0
requests==2.27.1
saga-tools @ git+https://github.com/isi-vista/saga-tools.git@b4c4d409f9b2f7e7bd578dc536fc30eb5de56f1c
scipy==1.7.3
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.3.2
sphinx-autodoc-typehints==1.16.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
temppathlib==1.1.0
toml==0.10.2
tomli==2.0.1
towncrier==21.3.0
typed-ast==1.5.2
types-PyYAML==6.0.4
types-tabulate==0.8.5
typing_extensions==4.0.1
urllib3==1.26.8
vistautils==0.24.0
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
